refactor: 타입 정의를 새로운 파일로 분리하고, 수신자 정보의 오타 수정
